chore: improved test coverage
- add public-path integration coverage for passwd, askpass login, crypto, and XML flows
- refactor password prompt, HTTP request setup, and crypto helpers into clearer units
- keep production behavior unchanged while making coverage gaps easier to isolate
