+++ src/crypto.rs
@@ -102,46 +102,16 @@ pub fn aes_decrypt_lastpass(ciphertext: &[u8], key: &[u8; 32]) -> Result<Vec<u8>
     }
 
     if ciphertext[0] == b'!' {
-        if ciphertext.len() < 1 + IV_LEN {
-            return Err(LpassError::Crypto("ciphertext too short"));
-        }
-        let iv = &ciphertext[1..1 + IV_LEN];
-        let data = &ciphertext[1 + IV_LEN..];
-
-        let mut buffer = data.to_vec();
-        let plaintext = Aes256CbcDec::new(key.into(), iv.into())
-            .decrypt_padded_mut::<Pkcs7>(&mut buffer)
-            .map_err(|_| LpassError::Crypto("AES-CBC decrypt failed"))?
-            .to_vec();
-        Ok(plaintext)
-    } else {
-        use cipher::block_padding::Pkcs7 as EcbPkcs7;
-        use cipher::{BlockDecryptMut as _, KeyInit as _};
-        use ecb::Decryptor as EcbDecryptor;
-
-        let mut buffer = ciphertext.to_vec();
-        let plaintext = EcbDecryptor::<Aes256>::new(key.into())
-            .decrypt_padded_mut::<EcbPkcs7>(&mut buffer)
-            .map_err(|_| LpassError::Crypto("AES-ECB decrypt failed"))?
-            .to_vec();
-        Ok(plaintext)
+        return aes_decrypt_lastpass_cbc(ciphertext, key);
     }
+    aes_decrypt_lastpass_legacy_ecb(ciphertext, key)
 }
 
 pub fn decrypt_private_key(private_key_enc: &str, key: &[u8; 32]) -> Result<Vec<u8>> {
     let decrypted = if private_key_enc.starts_with('!') {
         aes_decrypt_base64_lastpass(private_key_enc, key)?
     } else {
-        if !private_key_enc.len().is_multiple_of(2) {
-            return Err(LpassError::Crypto("invalid private key format"));
-        }
-        let mut encrypted = Vec::with_capacity(1 + 16 + private_key_enc.len() / 2);
-        encrypted.push(b'!');
-        encrypted.extend_from_slice(&key[..16]);
-        let hex_bytes =
-            hex::decode(private_key_enc).map_err(|_| LpassError::Crypto("invalid private key"))?;
-        encrypted.extend_from_slice(&hex_bytes);
-        aes_decrypt_lastpass(&encrypted, key)?
+        decrypt_legacy_private_key_payload(private_key_enc, key)?
     };
 
     let text = String::from_utf8_lossy(&decrypted);
@@ -189,12 +159,58 @@ pub fn rsa_decrypt_oaep(private_key_der: &[u8], ciphertext: &[u8]) -> Result<Vec
 }
 
 pub fn rsa_encrypt_oaep(public_key_der: &[u8], plaintext: &[u8]) -> Result<Vec<u8>> {
-    let public_key = RsaPublicKey::from_pkcs1_der(public_key_der)
-        .or_else(|_| RsaPublicKey::from_public_key_der(public_key_der))
-        .map_err(|_| LpassError::Crypto("invalid rsa public key"))?;
-    public_key
+    let public_key = rsa_public_key_from_der(public_key_der)?;
+    let ciphertext = public_key
         .encrypt(&mut OsRng, Oaep::new::<Sha1>(), plaintext)
-        .map_err(|_| LpassError::Crypto("RSA encrypt failed"))
+        .map_err(|_| LpassError::Crypto("RSA encrypt failed"))?;
+    Ok(ciphertext)
+}
+
+fn aes_decrypt_lastpass_cbc(ciphertext: &[u8], key: &[u8; 32]) -> Result<Vec<u8>> {
+    if ciphertext.len() < 1 + IV_LEN {
+        return Err(LpassError::Crypto("ciphertext too short"));
+    }
+    let iv = &ciphertext[1..1 + IV_LEN];
+    let data = &ciphertext[1 + IV_LEN..];
+    let mut buffer = data.to_vec();
+    let plaintext = Aes256CbcDec::new(key.into(), iv.into())
+        .decrypt_padded_mut::<Pkcs7>(&mut buffer)
+        .map_err(|_| LpassError::Crypto("AES-CBC decrypt failed"))?
+        .to_vec();
+    Ok(plaintext)
+}
+
+fn aes_decrypt_lastpass_legacy_ecb(ciphertext: &[u8], key: &[u8; 32]) -> Result<Vec<u8>> {
+    use cipher::block_padding::Pkcs7 as EcbPkcs7;
+    use cipher::{BlockDecryptMut as _, KeyInit as _};
+    use ecb::Decryptor as EcbDecryptor;
+
+    let mut buffer = ciphertext.to_vec();
+    let plaintext = EcbDecryptor::<Aes256>::new(key.into())
+        .decrypt_padded_mut::<EcbPkcs7>(&mut buffer)
+        .map_err(|_| LpassError::Crypto("AES-ECB decrypt failed"))?;
+    Ok(plaintext.to_vec())
+}
+
+fn decrypt_legacy_private_key_payload(private_key_enc: &str, key: &[u8; 32]) -> Result<Vec<u8>> {
+    if !private_key_enc.len().is_multiple_of(2) {
+        return Err(LpassError::Crypto("invalid private key format"));
+    }
+    let mut encrypted = Vec::with_capacity(1 + 16 + private_key_enc.len() / 2);
+    encrypted.push(b'!');
+    encrypted.extend_from_slice(&key[..16]);
+    encrypted.extend_from_slice(&decode_private_key_hex(private_key_enc)?);
+    aes_decrypt_lastpass(&encrypted, key)
+}
+
+fn decode_private_key_hex(private_key_enc: &str) -> Result<Vec<u8>> {
+    hex::decode(private_key_enc).map_err(|_| LpassError::Crypto("invalid private key"))
+}
+
+fn rsa_public_key_from_der(public_key_der: &[u8]) -> Result<RsaPublicKey> {
+    RsaPublicKey::from_pkcs1_der(public_key_der)
+        .or_else(|_| RsaPublicKey::from_public_key_der(public_key_der))
+        .map_err(|_| LpassError::Crypto("invalid rsa public key"))
 }
 
 pub fn sha256_hex(bytes: &[u8]) -> String {
@@ -245,129 +261,10 @@ pub fn base64_lastpass_decode(s: &str) -> Result<Vec<u8>> {
 
 pub fn aes_decrypt_base64_lastpass(ciphertext_b64: &str, key: &[u8; 32]) -> Result<Vec<u8>> {
     let bytes = base64_lastpass_decode(ciphertext_b64)?;
-    aes_decrypt_lastpass(&bytes, key)
+    let plaintext = aes_decrypt_lastpass(&bytes, key)?;
+    Ok(plaintext)
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn encrypt_decrypt_roundtrip() {
-        let key = [7u8; 32];
-        let plaintext = b"secret payload";
-        let encrypted = encrypt_authenticated(&key, plaintext).expect("encrypt");
-        let decrypted = decrypt_authenticated(&key, &encrypted).expect("decrypt");
-        assert_eq!(decrypted, plaintext);
-    }
-
-    #[test]
-    fn decrypt_rejects_tampered_data() {
-        let key = [42u8; 32];
-        let plaintext = b"secret payload";
-        let mut encrypted = encrypt_authenticated(&key, plaintext).expect("encrypt");
-        let last = encrypted.len() - 1;
-        encrypted[last] ^= 0x55;
-        let err = decrypt_authenticated(&key, &encrypted).expect_err("should fail");
-        match err {
-            LpassError::Crypto(_) => {}
-            other => panic!("unexpected error: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn base64_lastpass_roundtrip() {
-        let key = [1u8; 32];
-        let plaintext = b"hello";
-        let encrypted = aes_encrypt_lastpass(plaintext, &key).expect("encrypt");
-        let encoded = base64_lastpass_encode(&encrypted);
-        let decoded = base64_lastpass_decode(&encoded).expect("decode");
-        let decrypted = aes_decrypt_lastpass(&decoded, &key).expect("decrypt");
-        assert_eq!(decrypted, plaintext);
-    }
-
-    #[test]
-    fn decrypt_authenticated_rejects_short_ciphertext() {
-        let key = [7u8; 32];
-        let err = decrypt_authenticated(&key, b"short").expect_err("must fail");
-        assert!(matches!(err, LpassError::Crypto("ciphertext too short")));
-    }
-
-    #[test]
-    fn aes_lastpass_encrypt_decrypt_and_errors() {
-        let key = [9u8; 32];
-        let plaintext = b"payload";
-        let encrypted = aes_encrypt_lastpass(plaintext, &key).expect("encrypt");
-        let decrypted = aes_decrypt_lastpass(&encrypted, &key).expect("decrypt");
-        assert_eq!(decrypted, plaintext);
-
-        let err = aes_decrypt_lastpass(b"", &key).expect_err("empty must fail");
-        assert!(matches!(err, LpassError::Crypto("ciphertext empty")));
-
-        let err = aes_decrypt_lastpass(b"!", &key).expect_err("short must fail");
-        assert!(matches!(err, LpassError::Crypto("ciphertext too short")));
-    }
-
-    #[test]
-    fn base64_lastpass_decode_rejects_invalid_inputs() {
-        let err = base64_lastpass_decode("").expect_err("empty must fail");
-        assert!(matches!(err, LpassError::Crypto("base64 empty")));
-
-        let err = base64_lastpass_decode("!abc").expect_err("format must fail");
-        assert!(matches!(err, LpassError::Crypto("invalid base64 format")));
-
-        let err = base64_lastpass_decode("!!!|!!!").expect_err("decode must fail");
-        assert!(matches!(err, LpassError::Crypto("base64 decode failed")));
-    }
-
-    #[test]
-    fn aes_decrypt_base64_lastpass_roundtrip() {
-        let key = [11u8; 32];
-        let encrypted = aes_encrypt_lastpass(b"xyz", &key).expect("encrypt");
-        let encoded = base64_lastpass_encode(&encrypted);
-        let decrypted = aes_decrypt_base64_lastpass(&encoded, &key).expect("decrypt");
-        assert_eq!(decrypted, b"xyz");
-    }
-
-    #[test]
-    fn decrypt_private_key_rejects_invalid_formats() {
-        let key = [13u8; 32];
-        let err = decrypt_private_key("abc", &key).expect_err("odd hex must fail");
-        assert!(matches!(
-            err,
-            LpassError::Crypto("invalid private key format")
-        ));
-
-        let err = decrypt_private_key("zz", &key).expect_err("invalid hex must fail");
-        assert!(matches!(err, LpassError::Crypto("invalid private key")));
-    }
-
-    #[test]
-    fn encrypt_private_key_roundtrip() {
-        let key = [15u8; 32];
-        let private_key = vec![0x30, 0x82, 0x01, 0x0a, 0xde, 0xad, 0xbe, 0xef];
-        let encrypted = encrypt_private_key(&private_key, &key).expect("encrypt");
-        let decrypted = decrypt_private_key(&encrypted, &key).expect("decrypt");
-        assert_eq!(decrypted, private_key);
-    }
-
-    #[test]
-    fn rsa_decrypt_oaep_rejects_invalid_private_key() {
-        let err = rsa_decrypt_oaep(b"not-a-key", b"ciphertext").expect_err("must fail");
-        assert!(matches!(err, LpassError::Crypto("invalid rsa private key")));
-    }
-
-    #[test]
-    fn rsa_encrypt_oaep_rejects_invalid_public_key() {
-        let err = rsa_encrypt_oaep(b"not-a-key", b"ciphertext").expect_err("must fail");
-        assert!(matches!(err, LpassError::Crypto("invalid rsa public key")));
-    }
-
-    #[test]
-    fn sha256_hex_matches_reference() {
-        assert_eq!(
-            sha256_hex(b"abc"),
-            "ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f20015ad"
-        );
-    }
-}
+#[path = "crypto_tests.rs"]
+mod tests;
