+++ src/password.rs
@@ -1,5 +1,6 @@
 #![forbid(unsafe_code)]
 
+use std::io::Write;
 use std::process::Command;
 
 use crate::error::{LpassError, Result};
@@ -21,23 +22,34 @@ pub fn prompt_password_with_description(
     description: &str,
 ) -> Result<String> {
     if let Some(askpass) = askpass_program_from_env() {
-        let output = Command::new(askpass)
-            .arg(prompt)
-            .output()
-            .map_err(|err| LpassError::io("askpass", err))?;
-        if !output.status.success() {
-            return Err(LpassError::Crypto("askpass failed"));
-        }
-        return Ok(decode_password_output(output.stdout));
+        return prompt_password_with_askpass(&askpass, prompt);
     }
+    prompt_password_from_tty_with(prompt, error, description, rpassword::prompt_password)
+}
 
-    eprintln!("{}", description);
-    eprintln!();
-    if let Some(error) = error {
-        eprintln!("{}", error);
+fn prompt_password_with_askpass(askpass: &str, prompt: &str) -> Result<String> {
+    let output = Command::new(askpass)
+        .arg(prompt)
+        .output()
+        .map_err(|err| LpassError::io("askpass", err))?;
+    if !output.status.success() {
+        return Err(LpassError::Crypto("askpass failed"));
     }
-    rpassword::prompt_password(format!("{prompt}: "))
-        .map_err(|err| LpassError::io("password prompt", err))
+    Ok(decode_password_output(output.stdout))
+}
+
+fn prompt_password_from_tty_with<F>(
+    prompt: &str,
+    error: Option<&str>,
+    description: &str,
+    prompt_fn: F,
+) -> Result<String>
+where
+    F: FnOnce(String) -> std::io::Result<String>,
+{
+    write_prompt_description(&mut std::io::stderr().lock(), description, error)
+        .map_err(|err| LpassError::io("password prompt", err))?;
+    prompt_fn(format!("{prompt}: ")).map_err(|err| LpassError::io("password prompt", err))
 }
 
 fn askpass_program_from_env() -> Option<String> {
@@ -55,60 +67,19 @@ fn decode_password_output(mut bytes: Vec<u8>) -> String {
     String::from_utf8_lossy(&bytes).to_string()
 }
 
-#[cfg(test)]
-mod tests {
-    use super::{decode_password_output, prompt_password};
-    use tempfile::TempDir;
-
-    #[test]
-    fn decode_password_output_trims_trailing_newlines() {
-        let raw = b"passphrase with spaces\r\n".to_vec();
-        assert_eq!(decode_password_output(raw), "passphrase with spaces");
+fn write_prompt_description<W: Write>(
+    writer: &mut W,
+    description: &str,
+    error: Option<&str>,
+) -> std::io::Result<()> {
+    writeln!(writer, "{description}")?;
+    writeln!(writer)?;
+    if let Some(error) = error {
+        writeln!(writer, "{error}")?;
     }
-
-    #[test]
-    fn askpass_uses_non_empty_value() {
-        let value = super::askpass_program_from_value(Some("/tmp/lpass-askpass.sh".to_string()));
-        assert_eq!(value.as_deref(), Some("/tmp/lpass-askpass.sh"));
-    }
-
-    #[test]
-    fn askpass_ignores_empty_values() {
-        let value = super::askpass_program_from_value(Some("   ".to_string()));
-        assert!(value.is_none());
-    }
-
-    #[test]
-    #[cfg(unix)]
-    fn prompt_password_reads_value_from_askpass_program() {
-        use std::os::unix::fs::PermissionsExt;
-
-        let _guard = crate::lpenv::begin_test_overrides();
-        let temp = TempDir::new().expect("tempdir");
-        let script = temp.path().join("askpass-ok.sh");
-        std::fs::write(&script, "#!/bin/sh\necho from-askpass\n").expect("write script");
-        std::fs::set_permissions(&script, std::fs::Permissions::from_mode(0o700))
-            .expect("chmod script");
-        crate::lpenv::set_override_for_tests("LPASS_ASKPASS", &script.display().to_string());
-
-        let value = prompt_password("user@example.com").expect("askpass value");
-        assert_eq!(value, "from-askpass");
-    }
-
-    #[test]
-    #[cfg(unix)]
-    fn prompt_password_reports_askpass_failure() {
-        use std::os::unix::fs::PermissionsExt;
-
-        let _guard = crate::lpenv::begin_test_overrides();
-        let temp = TempDir::new().expect("tempdir");
-        let script = temp.path().join("askpass-fail.sh");
-        std::fs::write(&script, "#!/bin/sh\nexit 1\n").expect("write script");
-        std::fs::set_permissions(&script, std::fs::Permissions::from_mode(0o700))
-            .expect("chmod script");
-        crate::lpenv::set_override_for_tests("LPASS_ASKPASS", &script.display().to_string());
-
-        let err = prompt_password("user@example.com").expect_err("askpass must fail");
-        assert!(format!("{err}").contains("askpass failed"));
-    }
+    writer.flush()
 }
+
+#[cfg(test)]
+#[path = "password_tests.rs"]
+mod tests;
