+++ src/xml.rs
@@ -103,8 +103,12 @@ pub fn parse_error_cause(xml: &str, attr_name: &str) -> Option<String> {
 }
 
 pub fn parse_lastpass_api_ok(xml: &str) -> Option<bool> {
-    let attrs = parse_element_attributes(xml, b"lastpass")?;
-    Some(attrs.get("rc").map(String::as_str) == Some("OK"))
+    Some(
+        parse_element_attributes(xml, b"lastpass")?
+            .get("rc")
+            .map(String::as_str)
+            == Some("OK"),
+    )
 }
 
 pub fn parse_pwchange(xml: &str) -> std::result::Result<PwChangeInfo, PwChangeParseError> {
@@ -175,7 +179,11 @@ fn parse_pwchange_data(
         });
     }
 
-    Ok((reencrypt_id.to_string(), privkey_encrypted.to_string(), fields))
+    Ok((
+        reencrypt_id.to_string(),
+        privkey_encrypted.to_string(),
+        fields,
+    ))
 }
 
 fn split_pwchange_line(data: &str) -> Option<(&str, &str)> {
@@ -183,9 +191,7 @@ fn split_pwchange_line(data: &str) -> Option<(&str, &str)> {
     Some((&data[..newline], &data[newline + 1..]))
 }
 
-fn parse_pwchange_su_keys(
-    attrs: &std::collections::HashMap<String, String>,
-) -> Vec<PwChangeSuKey> {
+fn parse_pwchange_su_keys(attrs: &std::collections::HashMap<String, String>) -> Vec<PwChangeSuKey> {
     let mut su_keys = Vec::new();
     for idx in 0.. {
         let sukey = format!("sukey{idx}");
@@ -240,107 +246,5 @@ fn parse_element_attributes(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn parse_ok_session_basic() {
-        let xml = "<response><ok uid=\"57747756\" sessionid=\"1234\" token=\"abcd\"/></response>";
-        let session = parse_ok_session(xml).expect("session");
-        assert_eq!(session.uid, "57747756");
-        assert_eq!(session.session_id, "1234");
-        assert_eq!(session.token, "abcd");
-    }
-
-    #[test]
-    fn parse_ok_session_privatekey_fallback() {
-        let xml =
-            "<response><ok uid=\"1\" sessionid=\"2\" token=\"3\" privatekey=\"enc\"/></response>";
-        let session = parse_ok_session(xml).expect("session");
-        assert_eq!(session.private_key_enc.as_deref(), Some("enc"));
-    }
-
-    #[test]
-    fn parse_error_attribute_message() {
-        let xml = "<response><error message=\"invalid password\"/></response>";
-        let msg = parse_error_cause(xml, "message").expect("message");
-        assert_eq!(msg, "invalid password");
-    }
-
-    #[test]
-    fn parse_login_check_version() {
-        let xml = "<response><ok uid=\"1\" sessionid=\"2\" token=\"3\" accts_version=\"123\" url_encryption=\"1\" url_logging=\"1\"/></response>";
-        let mut session = Session::default();
-        let version = parse_login_check(xml, &mut session).expect("version");
-        assert_eq!(version, 123);
-        assert_eq!(session.uid, "1");
-        assert!(session.url_encryption_enabled);
-        assert!(session.url_logging_enabled);
-    }
-
-    #[test]
-    fn parse_ok_session_feature_flags() {
-        let xml = "<response><ok uid=\"1\" sessionid=\"2\" token=\"3\" url_encryption=\"1\" url_logging=\"0\"/></response>";
-        let session = parse_ok_session(xml).expect("session");
-        assert!(session.url_encryption_enabled);
-        assert!(!session.url_logging_enabled);
-    }
-
-    #[test]
-    fn parse_ok_session_returns_none_for_missing_required_attributes() {
-        assert!(parse_ok_session("<response><ok uid=\"1\" sessionid=\"2\"/></response>").is_none());
-    }
-
-    #[test]
-    fn parse_lastpass_api_ok_status() {
-        assert_eq!(
-            parse_lastpass_api_ok("<lastpass rc=\"OK\"><ok/></lastpass>"),
-            Some(true)
-        );
-        assert_eq!(
-            parse_lastpass_api_ok("<lastpass rc=\"FAIL\"><error/></lastpass>"),
-            Some(false)
-        );
-        assert_eq!(parse_lastpass_api_ok("<response/>"), None);
-    }
-
-    #[test]
-    fn parse_error_cause_returns_none_for_malformed_xml() {
-        assert_eq!(parse_error_cause("<response><error", "message"), None);
-    }
-
-    #[test]
-    fn parse_pwchange_success() {
-        let xml = "<lastpass rc=\"OK\"><data token=\"tok\" sukey0=\"0102\" suuid0=\"77\" xml=\"rid&#10;priv&#10;old-a&#10;old-b&#9;0&#10;endmarker&#10;\"/></lastpass>";
-        let info = parse_pwchange(xml).expect("pwchange");
-        assert_eq!(info.reencrypt_id, "rid");
-        assert_eq!(info.token, "tok");
-        assert_eq!(info.privkey_encrypted, "priv");
-        assert_eq!(info.fields.len(), 2);
-        assert!(!info.fields[0].optional);
-        assert!(info.fields[1].optional);
-        assert_eq!(info.su_keys.len(), 1);
-        assert_eq!(info.su_keys[0].uid, "77");
-        assert_eq!(info.su_keys[0].sharing_key, vec![1, 2]);
-    }
-
-    #[test]
-    fn parse_pwchange_rejects_incorrect_password() {
-        let err = parse_pwchange("<lastpass rc=\"FAIL\"><error/></lastpass>").expect_err("fail");
-        assert_eq!(err, PwChangeParseError::IncorrectPassword);
-    }
-
-    #[test]
-    fn parse_pwchange_rejects_invalid_payload() {
-        let err = parse_pwchange("<lastpass rc=\"OK\"><data token=\"tok\" xml=\"rid\"/></lastpass>")
-            .expect_err("invalid");
-        assert_eq!(err, PwChangeParseError::Invalid);
-    }
-
-    #[test]
-    fn parse_pwchange_stops_su_key_scan_on_invalid_hex() {
-        let xml = "<lastpass rc=\"OK\"><data token=\"tok\" sukey0=\"not-hex\" suuid0=\"77\" sukey1=\"0102\" suuid1=\"88\" xml=\"rid&#10;priv&#10;endmarker&#10;\"/></lastpass>";
-        let info = parse_pwchange(xml).expect("pwchange");
-        assert!(info.su_keys.is_empty());
-    }
-}
+#[path = "xml_tests.rs"]
+mod tests;
